#193: move native smoke tests to platform.*, dedup pickFolder, cover empty-drop reset
