+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/transfer/MacFilePicker.kt
@@ -2,13 +2,11 @@ package com.tubetoast.tether.transfer
 
 import com.tubetoast.tether.platform.mac.Foundation
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.swing.Swing
 import kotlinx.coroutines.withContext
 import ru.pocketbyte.kydra.log.KydraLog
 import ru.pocketbyte.kydra.log.info
 import ru.pocketbyte.kydra.log.wrapper.withTag
 import java.io.File
-import javax.swing.JFileChooser
 
 private val log = KydraLog.withTag(default = "MacFilePicker")
 
@@ -23,21 +21,7 @@ internal class MacFilePicker(
         }
     }
 
-    override suspend fun pickFolder(): List<FileSource> {
-        val dir = withContext(Dispatchers.Swing) {
-            val chooser = JFileChooser()
-            chooser.fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-            val result = chooser.showOpenDialog(windowHolder.window)
-            if (result == JFileChooser.APPROVE_OPTION) chooser.selectedFile else null
-        }
-        if (dir == null) {
-            log.info { "pickFolder (mac): cancelled" }
-            return emptyList()
-        }
-        val sources = withContext(Dispatchers.IO) { JvmFolderWalker().walk(dir) }
-        log.info { "pickFolder (mac): ${sources.size} file(s) from ${dir.name}" }
-        return sources
-    }
+    override suspend fun pickFolder(): List<FileSource> = jvmPickFolder(windowHolder)
 
     override suspend fun pickPhotos(): List<FileSource> =
         throw UnsupportedOperationException("pickPhotos is mobile-only")
