+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/transfer/WindowsFilePicker.kt
@@ -21,7 +21,6 @@ import com.tubetoast.tether.platform.win.ShTypes.SIGDN.Companion.SIGDN_FILESYSPA
 import com.tubetoast.tether.platform.win.ShellItem
 import com.tubetoast.tether.platform.win.ShellItemArray
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.swing.Swing
 import kotlinx.coroutines.withContext
 import ru.pocketbyte.kydra.log.KydraLog
 import ru.pocketbyte.kydra.log.info
@@ -29,7 +28,6 @@ import ru.pocketbyte.kydra.log.warn
 import ru.pocketbyte.kydra.log.wrapper.withTag
 import java.awt.Window
 import java.io.File
-import javax.swing.JFileChooser
 
 private val log = KydraLog.withTag(default = "WindowsFilePicker")
 
@@ -43,21 +41,7 @@ internal class WindowsFilePicker(
                 awtPickFiles(windowHolder)
             }
 
-    override suspend fun pickFolder(): List<FileSource> {
-        val dir = withContext(Dispatchers.Swing) {
-            val chooser = JFileChooser()
-            chooser.fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-            val result = chooser.showOpenDialog(windowHolder.window)
-            if (result == JFileChooser.APPROVE_OPTION) chooser.selectedFile else null
-        }
-        if (dir == null) {
-            log.info { "pickFolder (windows): cancelled" }
-            return emptyList()
-        }
-        val sources = withContext(Dispatchers.IO) { JvmFolderWalker().walk(dir) }
-        log.info { "pickFolder (windows): ${sources.size} file(s) from ${dir.name}" }
-        return sources
-    }
+    override suspend fun pickFolder(): List<FileSource> = jvmPickFolder(windowHolder)
 
     override suspend fun pickPhotos(): List<FileSource> =
         throw UnsupportedOperationException("pickPhotos is mobile-only")
